+++ src/fluoscenepy/tests/test_fluoscene.py
@@ -62,12 +62,15 @@ def test_fluorobj_initialization():
         pass
     flobj = FluorObj(typical_size=4.75); flobj.get_shape(); flobj.crop_shape()
     assert flobj.profile.shape[0] > 4 and flobj.profile.shape[1] > 4, f"Profile sizes out of the expected range (5, 5): {flobj.profile.shape}"
-    flobj = FluorObj(typical_size=(4.2, 5.1, 0.25*np.pi), shape_type='el'); flobj.get_shape()
+    flobj = FluorObj(typical_size=(4.2, 5.1, 0.25*np.pi), shape_type='el'); flobj.get_shape(accelerated=False)
     flobj.get_casted_shape(max_pixel_value=1921, image_type='uint16'); flobj.crop_shape(); flobj.crop_shape()
     assert flobj.profile.shape == flobj.casted_profile.shape, "Profile and casted profile shapes aren't equa or double cropping causes errors"
     flobj = FluorObj(typical_size=2.0); flobj.get_shape(accelerated=True)
     assert flobj.profile.shape[1] == 3 and flobj.profile.shape[0] == 3, ("Profile sizes out of the expected range (3, 3): "
                                                                          + f"{flobj.profile.shape}")
+    flobj = FluorObj(typical_size=(3.2, 4.2, -0.51*np.pi), shape_type='el'); flobj.get_shape(accelerated=True); flobj.crop_shape()
+    flobj.get_casted_shape(max_pixel_value=100.0, image_type=np.float64)
+    assert np.max(flobj.casted_profile) >= 100.0, f"Profile not casted properly, max value: {np.max(flobj.profile)}"
 
 
 def test_objects_generation():