+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
-Logging of changes between package versions (generated and uploaded to pypi.org).
-All notable changes to this project will be documented in this file.    
+Logging of changes between package versions (generated and uploaded to pypi.org).     
+All notable changes to this project will be documented in this file.       
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).   
 
-## [0.0.1] - 2024-19-09
+## [0.0.2] - 2024-12-...     
+Overall acceleration of computation and object placement methods.
+### Added
+- Methods (check API documentation): ...
+- More printouts if methods for *UscopeScene* called with *verbose_info* flag.
+
+## [0.0.1] - 2024-09-19
 Initial release with 2 implemented simulation classes:    
 *UscopeScene* - for simulation of a microscopic image;      
 *FluorObj* - for  simulation of an individual bead / elongated small continuously-shaped fluorescent object;